:truck:(chore) rename errors folder to exceptions
