+++ server/src/routes/v1/appointments.routes.ts
@@ -16,22 +16,18 @@ routes.get('/', async (request, response) => {
 });
 
 routes.post('/', async (request, response) => {
-  try {
-    const { provider_id, date } = request.body;
+  const { provider_id, date } = request.body;
 
-    const parsedDate = parseISO(date);
+  const parsedDate = parseISO(date);
 
-    const createAppointment = new CreateAppointmentService();
+  const createAppointment = new CreateAppointmentService();
 
-    const appointment = await createAppointment.execute({
-      provider_id,
-      date: parsedDate,
-    });
+  const appointment = await createAppointment.execute({
+    provider_id,
+    date: parsedDate,
+  });
 
-    return response.json(appointment);
-  } catch ({ status, message }) {
-    return response.status(status).json({ message });
-  }
+  return response.json(appointment);
 });
 
 export default routes;
