+++ server/src/routes/v1/sessions.routes.ts
@@ -5,22 +5,18 @@ import AuthenticateUserService from '../../services/AuthenticateUserService';
 const routes = Router();
 
 routes.post('/', async (request, response) => {
-  try {
-    const { email, password } = request.body;
+  const { email, password } = request.body;
 
-    const authenticateUser = new AuthenticateUserService();
+  const authenticateUser = new AuthenticateUserService();
 
-    const session = await authenticateUser.execute({
-      email,
-      password,
-    });
+  const session = await authenticateUser.execute({
+    email,
+    password,
+  });
 
-    const { user, token } = session;
+  const { user, token } = session;
 
-    return response.json(classToPlain({ user, token }));
-  } catch ({ status, message }) {
-    return response.status(status).json({ message });
-  }
+  return response.json(classToPlain({ user, token }));
 });
 
 export default routes;
