+++ server/src/routes/v1/users.routes.ts
@@ -10,21 +10,19 @@ const routes = Router();
 const upload = multer(uploadConfig);
 
 routes.post('/', async (request, response) => {
-  try {
-    const { name, email, password } = request.body;
+  const { name, email, password } = request.body;
 
-    const createUser = new CreateUserService();
+  const createUser = new CreateUserService();
 
-    const user = await createUser.execute({
-      name,
-      email,
-      password,
-    });
+  const user = await createUser.execute({
+    name,
+    email,
+    password,
+  });
 
-    return response.json(classToPlain(user));
-  } catch (err) {
-    return response.status(400).json({ message: err.message });
-  }
+  response.json();
+
+  return response.json(classToPlain(user));
 });
 
 routes.patch(
@@ -32,19 +30,15 @@ routes.patch(
   ensureAuthentication,
   upload.single('avatar'),
   async (request, response) => {
-    try {
-      const { user, file } = request;
-      const updateUserAvatar = new UpdateUserAvatarService();
+    const { user, file } = request;
+    const updateUserAvatar = new UpdateUserAvatarService();
 
-      const updatedUser = await updateUserAvatar.execute({
-        user_id: user.id,
-        filename: file.filename,
-      });
+    const updatedUser = await updateUserAvatar.execute({
+      user_id: user.id,
+      filename: file.filename,
+    });
 
-      return response.json(classToPlain(updatedUser));
-    } catch ({ status, message }) {
-      return response.status(status).json({ message });
-    }
+    return response.json(classToPlain(updatedUser));
   },
 );
